Refactor build process: replace Travis CI with multi-stage Docker builds and reorganize source files
