--- .travis.yml
@@ -1,8 +1,0 @@
-language: go
-go:
-  - 1.x
-script:
-  - ./test.sh
-notifications:
-  email: false
-sudo: false
