+++ src/cmd/version.go
@@ -1,3 +1,3 @@
-package main
+package cmd
 
 const VERSION = "1.0.0"
